refactor(mobile): split category sheet
